Add GitHub OAuth config
