+++ lib/ci-config-stack.ts
@@ -21,7 +21,7 @@ export class CIConfigStack extends Stack {
 
   static readonly REDIRECT_URL_SECRET_EXPORT_VALUE: string = 'redirectUrlSecret';
 
-  static readonly OIDC_CONFIGURATION_VALUE_SECRET_EXPORT_VALUE: string = 'OIDCConfigValueSecret';
+  static readonly AUTH_CONFIGURATION_VALUE_SECRET_EXPORT_VALUE: string = 'authConfigValueSecret';
 
   static readonly CASC_RELOAD_TOKEN_SECRET_EXPORT_VALUE: string = 'casc';
 
@@ -43,9 +43,10 @@ export class CIConfigStack extends Stack {
     const redirectUrlSecret = new Secret(this, 'redirectUrl', {
       description: 'Redirect url for Jenkins',
     });
-    const OIDCConfigValuesSecret = new Secret(this, 'OIDCConfigValues', {
-      description: 'OIDC params in JSON format',
+    const authConfigValuesSecret = new Secret(this, 'authConfigValues', {
+      description: 'Auth credentials in JSON format',
     });
+
     const CascReloadTokenValuesSecret = new Secret(this, 'CascReloadTokenValue', {
       description: 'Reload token (password) required for configuration as code plugin',
     });
@@ -75,9 +76,9 @@ export class CIConfigStack extends Stack {
       exportName: CIConfigStack.REDIRECT_URL_SECRET_EXPORT_VALUE,
     });
 
-    new CfnOutput(this, 'OIDCConfigValuesSecret', {
-      value: OIDCConfigValuesSecret.secretArn,
-      exportName: CIConfigStack.OIDC_CONFIGURATION_VALUE_SECRET_EXPORT_VALUE,
+    new CfnOutput(this, 'authConfigValuesSecret', {
+      value: authConfigValuesSecret.secretArn,
+      exportName: CIConfigStack.AUTH_CONFIGURATION_VALUE_SECRET_EXPORT_VALUE,
     });
 
     new CfnOutput(this, 'cascSecretValue', {
