+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "opensearch-ci-stack",
-  "version": "0.3.0",
+  "version": "1.0.0",
   "bin": {
     "ci": "bin/ci-stack.js"
   },
